+++ src/App.css
@@ -1,9 +1,12 @@
 * {
   box-sizing: border-box;
   font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+  padding: 0;
+  margin: 0;
 }
 
-.todo-app {
+.contacts-app {
+  margin: 5%;
   display: flex;
   align-items: center;
   justify-content: center;
@@ -32,11 +35,11 @@ button {
   cursor: pointer;
 }
 
-.todo-list {
+.contacts {
   min-width: 40%;
 }
 
-.todo-list li {
+.info {
   font-size: 1.2rem;
   list-style: none;
   border-radius: 5px;
@@ -46,7 +49,7 @@ button {
   cursor: pointer;
 } 
 
-.todo-list li:hover {
+.info:hover {
   background: #f1f1f1;
 }
